style: format the tree, and fix what the static analysers found
Preparing a CI gate for these tools meant running them for the first time.

.clang-format had two CommentPragmas keys, which makes the whole file invalid
to clang-format 17 and newer -- so the project's style has not actually been
enforceable for some time, and 26 of 34 files had drifted. The duplicate is
gone and the tree is formatted with clang-format 18, which is the version CI
pins; leaving it unpinned would turn a runner image update into a spurious
failure.

clang-tidy found one real defect: a dead store in copy_payload's partial-copy
fallback. dst_offset was advanced and never read again, because the remaining
bytes are addressed through the mapping while the source offset stays absolute.
The store is gone and the reason is written down.

The rest of what it found was smaller: multiplications performed in int before
widening to int64_t (harmless at these values, wrong in principle), three
variables that could be const, and a parameter passed by value that is only
copied. Two findings are deliberate and now say so at the site rather than
being disabled globally -- the empty catch in record_error, which exists so a
failure to record cannot escape a worker thread, and the memcpy into the name
blob, which is length-prefixed and so intentionally not NUL-terminated.

cppcheck could not parse main()'s function-try-block, which is valid C++ that
several analysers stumble over; an inner try is equivalent and costs nothing.
Its remaining complaint is a parser limitation on function_ref and is
suppressed by name, so a genuine syntax error still fails the run.

Both tools are configured explicitly rather than by default, so a new release
cannot silently change what CI enforces. performance-enum-size is off: the base
type of a public enum is an ABI decision, not a lint.

src/test.cpp was the version 1 benchmark, replaced by src/bench.cpp and unused
since; nothing built it.

Verified after formatting: gcc 12 and 14 and clang 19 and 22, each at C++20,
C++23 and C++26, plus gcc 15.3 and 16.1, Alpine musl on x86_64 and arm64.

+++ src/format.hpp
@@ -6,6 +6,7 @@
 #include <cstring>
 #include <stdexcept>
 #include <string>
+
 #include <type_traits>
 
 // uvfs format, version 2
@@ -91,16 +92,20 @@ static constexpr auto round_up(int64_t x, int64_t multiple) noexcept -> int64_t
   return static_cast<int64_t>(((static_cast<uint64_t>(x) + m - 1) / m) * m);
 }
 
+// The arithmetic is done unsigned so that rounding cannot overflow into
+// undefined behaviour; the result is back in range by construction.
+// NOLINTNEXTLINE(bugprone-narrowing-conversions)
 [[nodiscard]]
 static constexpr auto round_up_8(int64_t x) noexcept -> int64_t
 {
-  return ((static_cast<uint64_t>(x) + 7) >> 3) << 3;
+  return static_cast<int64_t>(((static_cast<uint64_t>(x) + 7) >> 3) << 3);
 }
 
+// NOLINTNEXTLINE(bugprone-narrowing-conversions)
 [[nodiscard]]
 static constexpr auto round_up_64(int64_t x) noexcept -> int64_t
 {
-  return ((static_cast<uint64_t>(x) + 63) >> 6) << 6;
+  return static_cast<int64_t>(((static_cast<uint64_t>(x) + 63) >> 6) << 6);
 }
 
 //! Smallest power of two that keeps the load factor at or below 70%.
@@ -115,7 +120,7 @@ static constexpr auto table_capacity_for(int64_t n) noexcept -> int64_t
   return cap;
 }
 
-template<typename T>
+template <typename T>
 static constexpr int64_t ssizeof = static_cast<int64_t>(sizeof(T));
 
 // Reads a scalar out of the mapping without assuming anything about the
@@ -123,7 +128,7 @@ static constexpr int64_t ssizeof = static_cast<int64_t>(sizeof(T));
 // offset, and forming a misaligned pointer there is undefined behaviour even
 // on architectures where the load itself would have worked. Compilers lower
 // this to a plain load when the target allows it.
-template<typename T>
+template <typename T>
 [[nodiscard]] inline auto load(const char* p) noexcept -> T
 {
   static_assert(std::is_trivially_copyable_v<T>);
@@ -132,7 +137,7 @@ template<typename T>
   return v;
 }
 
-template<typename T>
+template <typename T>
 inline void store(char* p, T v) noexcept
 {
   static_assert(std::is_trivially_copyable_v<T>);
@@ -142,9 +147,9 @@ inline void store(char* p, T v) noexcept
 //! One index entry. Fixed stride so the array is directly indexable.
 struct entry
 {
-  int64_t data_offset{}; //!< relative to header::data_start
-  int64_t stored_size{}; //!< bytes occupied in the archive
-  int64_t orig_size{};   //!< bytes after decompression
+  int64_t data_offset{};  //!< relative to header::data_start
+  int64_t stored_size{};  //!< bytes occupied in the archive
+  int64_t orig_size{};    //!< bytes after decompression
   uint32_t name_offset{}; //!< offset into the name blob
   uint16_t name_size{};
   codec method{codec::store};
@@ -257,10 +262,7 @@ struct header
   {
     return file_count * entry_size;
   }
-  [[nodiscard]] auto table_offset() const noexcept -> int64_t
-  {
-    return entries_bytes();
-  }
+  [[nodiscard]] auto table_offset() const noexcept -> int64_t { return entries_bytes(); }
   [[nodiscard]] auto table_bytes() const noexcept -> int64_t
   {
     return table_capacity * 8;
@@ -297,8 +299,7 @@ struct header
     if (version() != format_version)
       throw std::runtime_error(
           "uvfs: unsupported format version " + std::to_string(version())
-          + " (this build reads version " + std::to_string(format_version)
-          + "): ");
+          + " (this build reads version " + std::to_string(format_version) + "): ");
   }
 
   void validate(int64_t filesize) const
@@ -345,8 +346,8 @@ struct header
 
     // The index sub-regions must fit inside the index, in order.
     // file_count is bounded above, so these products cannot overflow.
-    const auto need = static_cast<uint64_t>(names_offset())
-                      + static_cast<uint64_t>(names_size);
+    const auto need
+        = static_cast<uint64_t>(names_offset()) + static_cast<uint64_t>(names_size);
     if (need > idx_sz)
       throw std::runtime_error(
           "uvfs: index is too small for its entries, table and names: ");
@@ -373,8 +374,7 @@ struct header
     }
     else if (dict_size != 0)
     {
-      throw std::runtime_error(
-          "uvfs: dictionary present but the flag is not set: ");
+      throw std::runtime_error("uvfs: dictionary present but the flag is not set: ");
     }
   }
 };
